refactor platform parameters
