+++ inc/Platform/Parameters/HousekeepingService.hpp
@@ -1,13 +1,13 @@
 #ifndef ECSS_SERVICES_HOUSEKEEPINGPARAMETERS_HPP
 #define ECSS_SERVICES_HOUSEKEEPINGPARAMETERS_HPP
 
-#include "AcubeSATParameters.hpp"
+#include "PeakSatParameters.hpp"
 #include "Services/HousekeepingService.hpp"
 
 namespace HousekeepingStructures {
-    using namespace AcubeSATParameters;
+    using namespace PeakSatParameters;
 
-    static etl::array <HousekeepingStructure, ECSSMaxHousekeepingStructures> structures = {
+    static etl::array<HousekeepingStructure, ECSSMaxHousekeepingStructures> structures = {
             {
                     // OBC Parameters that will be monitored in the campaign
                     {0, 1000, true,
@@ -57,43 +57,7 @@ namespace HousekeepingStructures {
                       OBCNANDFLASHScrubbingFrequency,
                       OBCRAMScrubbingFrequency,
                       OBCProgramFlashScrubbingFrequency}},
-                    // ADCS Parameters that will be monitored every 0.01 sec in the campaign
-                    {3, 10, false,
-                     {ADCSMagnetometerRawX,
-                      ADCSMagnetometerRawY,
-                      ADCSMagnetometerRawZ,
-                      ADCSGyroscopeRateX,
-                      ADCSGyroscopeRateY,
-                      ADCSGyroscopeRateZ}},
-                    // ADCS Parameters that will be monitored every 3 seconds in the campaign
-                    {4, 3000, true,
-                     {ADCSMagnetometerFrequency,
-                      ADCSMagnetometerCycleCountX,
-                      ADCSMagnetometerCycleCountY,
-                      ADCSMagnetometerCycleCountZ,
-                      ADCSMagnetometerSelfTest,
-                      ADCSGyroscopeXTemperature,
-                      ADCSGyroscopeYTemperature,
-                      ADCSGyroscopeZTemperature,
-                      ADCSBoardTemperature1,
-                      ADCSBoardTemperature2,
-                      ADCSMCUTemperature,
-                      ADCSBootCounter,
-                      ADCSOnBoardTime,
-                      ADCSSystick}},
-                    // ADCS Parameters that might not be used
-                    {5, 5000, false,
-                     {ADCSMagnetometerSignX,
-                      ADCSMagnetometerSignY,
-                      ADCSMagnetometerSignZ,
-                      ADCSGyroscopeRateX,
-                      ADCSGyroscopeRateY,
-                      ADCSGyroscopeRateZ,
-                      ADCSGyroBiasX,
-                      ADCSGyroBiasY,
-                      ADCSGyroBiasZ,
-                      ADCSFLASHInt,
-                      ADCSSRAMInt}},
+
             }
     };
 }
