+++ src/Platform/Services/ParameterService.cpp
@@ -3,421 +3,139 @@
 #ifdef SERVICE_PARAMETER
 
 #include "Services/ParameterService.hpp"
-#include "AcubeSATParameters.hpp"
+#include "PeakSatParameters.hpp"
 
 void ParameterService::initializeParameterMap() {
-    parameters = {{AcubeSATParameters::OBCMCUTemperature,                                    AcubeSATParameters::obcMCUTemperature},
-                  {AcubeSATParameters::OBCUseRTT,                                            AcubeSATParameters::obcUseRTT},
-                  {AcubeSATParameters::OBCUseUART,                                           AcubeSATParameters::obcUseUART},
-                  {AcubeSATParameters::OBCUseCAN,                                            AcubeSATParameters::obcUseCAN},
+    parameters = {{PeakSatParameters::OBCMCUTemperature,                                    PeakSatParameters::obcMCUTemperature},
+                  {PeakSatParameters::OBCUseRTT,                                            PeakSatParameters::obcUseRTT},
+                  {PeakSatParameters::OBCUseUART,                                           PeakSatParameters::obcUseUART},
+                  {PeakSatParameters::OBCUseCAN,                                            PeakSatParameters::obcUseCAN},
             //-----------------------------------------------------------------------------------------------
-                  {AcubeSATParameters::OBCPCBTemperature1,                                   AcubeSATParameters::obcPCBTemperature1},
-                  {AcubeSATParameters::OBCPCBTemperature2,                                   AcubeSATParameters::obcPCBTemperature2},
-                  {AcubeSATParameters::OBCMCUInputVoltage,                                   AcubeSATParameters::obcMCUInputVoltage},
-                  {AcubeSATParameters::OBCMCUBootCounter,                                    AcubeSATParameters::obcMCUBootCounter},
-                  {AcubeSATParameters::OBCFlashInt,                                          AcubeSATParameters::obcFlashInt},
-                  {AcubeSATParameters::OBCSRAMInt,                                           AcubeSATParameters::obcSRAMInt},
-                  {AcubeSATParameters::OBCAvailableMRAM,                                     AcubeSATParameters::obcAvailableMRAM},
-                  {AcubeSATParameters::OBCAvailableNAND,                                     AcubeSATParameters::obcAvailableNAND},
-                  {AcubeSATParameters::OBCSpacecraftTimeRef,                                 AcubeSATParameters::obcSpacecraftTimeRef},
-                  {AcubeSATParameters::OBCOnBoardTime,                                       AcubeSATParameters::obcOnBoardTime},
-                  {AcubeSATParameters::OBCOperationalMode,                                   AcubeSATParameters::obcOperationalMode},
-                  {AcubeSATParameters::OBCMemoryPartition,                                   AcubeSATParameters::obcMemoryPartition},
-                  {AcubeSATParameters::OBCReconfigurationTimer,                              AcubeSATParameters::obcReconfigurationTimer},
-                  {AcubeSATParameters::OBCLastFailedEvent,                                   AcubeSATParameters::obcLastFailedEvent},
-                  {AcubeSATParameters::OBCMCUSystick,                                        AcubeSATParameters::obcMCUSystick},
-                  {AcubeSATParameters::OBCCANBUSLoad1,                                       AcubeSATParameters::obcCANBUSLoad1},
-                  {AcubeSATParameters::OBCCANBUSLoad2,                                       AcubeSATParameters::obcCANBUSLoad2},
-                  {AcubeSATParameters::OBCCANBUSActive,                                      AcubeSATParameters::obcCANBUSActive},
-                  {AcubeSATParameters::OBCMCUFDIR,                                           AcubeSATParameters::obcMCUFDIR},
-                  {AcubeSATParameters::OBCMCURestartSafeModeThreshold,                       AcubeSATParameters::obcMCURestartSafeModeThreshold},
-                  {AcubeSATParameters::OBCNANDFLASHLCLThreshold,                             AcubeSATParameters::obcNANDFLASHLCLThreshold},
-                  {AcubeSATParameters::OBCMRAMLCLThreshold,                                  AcubeSATParameters::obcMRAMLCLThreshold},
-                  {AcubeSATParameters::OBCNANDFLASHON,                                       AcubeSATParameters::obcNANDFLASHON},
-                  {AcubeSATParameters::OBCMRAMON,                                            AcubeSATParameters::obcNANDFlashScrubbingFrequency},
-                  {AcubeSATParameters::OBCNANDFLASHScrubbingFrequency,                       AcubeSATParameters::obcMRAMON},
-                  {AcubeSATParameters::OBCRAMScrubbingFrequency,                             AcubeSATParameters::obcRAMScrubbingFrequency},
-                  {AcubeSATParameters::OBCProgramFlashScrubbingFrequency,                    AcubeSATParameters::obcProgramFlashScrubbingFrequency},
-            //-----------------------------------------------------------------------------------------------
-                  {AcubeSATParameters::SUMCUBootCounter,                                     AcubeSATParameters::suMCUBootCounter},
-                  {AcubeSATParameters::SUMCUSysTick,                                         AcubeSATParameters::suMCUSysTick},
-                  {AcubeSATParameters::SUMCUTemperature,                                     AcubeSATParameters::suMCUTemperature},
-                  {AcubeSATParameters::SUMCUInputVoltage,                                    AcubeSATParameters::suMCUInputVoltage},
-                  {AcubeSATParameters::SUFlashIntUsedPercentage,                             AcubeSATParameters::suFlashIntUsedPercentage},
-                  {AcubeSATParameters::SUFlashUsedPercentage,                                AcubeSATParameters::suFlashUsedPercentage},
-                  {AcubeSATParameters::SUSRAMUsedPercentage,                                 AcubeSATParameters::suSRAMUsedPercentage},
-                  {AcubeSATParameters::SUNANDCurrentlyUsedPartition,                         AcubeSATParameters::suNANDCurrentlyUsedPartition},
-                  {AcubeSATParameters::SUNANDFlashLCLOn,                                     AcubeSATParameters::suNANDFlashLclOn},
-                  {AcubeSATParameters::SUNANDFlashLCLThreshold,                              AcubeSATParameters::suNANDFlashLclThreshold},
-                  {AcubeSATParameters::SUNANDFlashScrubbingFrequency,                        AcubeSATParameters::suNANDFlashScrubbingFrequency},
-                  {AcubeSATParameters::SURAMScrubbingFrequency,                              AcubeSATParameters::suRAMScrubbingFrequency},
-                  {AcubeSATParameters::SUProgramFlashScrubbingFrequency,                     AcubeSATParameters::suProgramFlashScrubbingFrequency},
-                  {AcubeSATParameters::SUControlPumpFlashLCLOn,                              AcubeSATParameters::suControlPumpFlashLclOn},
-                  {AcubeSATParameters::SUControlPumpFlashLCLThreshold,                       AcubeSATParameters::suControlPumpFlashLclThreshold},
-                  {AcubeSATParameters::SUFlowPumpFlashOn,                                    AcubeSATParameters::suFlowPumpFlashOn},
-                  {AcubeSATParameters::SUFlowPumpFlashLCLThreshold,                          AcubeSATParameters::suFlowPumpFlashLclThreshold},
-                  {AcubeSATParameters::SUCameraON,                                           AcubeSATParameters::suCameraOn},
-                  {AcubeSATParameters::SUCameraLCLThreshold,                                 AcubeSATParameters::suCameraLCLThreshold},
-                  {AcubeSATParameters::SUPCBTemperature1,                                    AcubeSATParameters::suPCBTemperature1},
-                  {AcubeSATParameters::SUPCBTemperature2,                                    AcubeSATParameters::suPCBTemperature2},
-                  {AcubeSATParameters::SUPCBTemperatureControlValves,                        AcubeSATParameters::suPCBTemperatureControlValves},
-                  {AcubeSATParameters::SUPCBTemperatureFlowValves,                           AcubeSATParameters::suPCBTemperatureFlowValves},
-                  {AcubeSATParameters::SUCameraTemperature,                                  AcubeSATParameters::suCameraTemperature},
-                  {AcubeSATParameters::SUPDMSTemperature,                                    AcubeSATParameters::suPDMSTemperature},
-                  {AcubeSATParameters::SUGrowthMediumTemperature,                            AcubeSATParameters::suGrowthMediumTemperature},
-                  {AcubeSATParameters::SULastFailedEvent,                                    AcubeSATParameters::suLastFailedEvent},
-                  {AcubeSATParameters::SUPressureLevelSensor1,                               AcubeSATParameters::suPressureLevelSensor1},
-                  {AcubeSATParameters::SUPressureLevelSensor2,                               AcubeSATParameters::suPressureLevelSensor2},
-                  {AcubeSATParameters::SUHumidityLevelSensor1,                               AcubeSATParameters::suHumidityLevelSensor1},
-                  {AcubeSATParameters::SUHumidityLevelSensor2,                               AcubeSATParameters::suHumidityLevelSensor2},
-                  {AcubeSATParameters::SURADFET,                                             AcubeSATParameters::suRadfet},
-                  {AcubeSATParameters::SUPumpFlowRate1,                                      AcubeSATParameters::suPumpFlowRate1},
-                  {AcubeSATParameters::SUPumpFlowRate2,                                      AcubeSATParameters::suPumpFlowRate2},
-                  {AcubeSATParameters::SUControlChamberValve1,                               AcubeSATParameters::suCcontrolChamberValve1},
-                  {AcubeSATParameters::SUControlChamberValve2,                               AcubeSATParameters::suControlChamberValve2},
-                  {AcubeSATParameters::SUControlChamberValve3,                               AcubeSATParameters::suControlChamberValve3},
-                  {AcubeSATParameters::SUControlSandwitchValve1,                             AcubeSATParameters::suControlSandwitchValve1},
-                  {AcubeSATParameters::SUControlSandwitchValve2,                             AcubeSATParameters::suControlSandwitchValve2},
-                  {AcubeSATParameters::SUControlSandwitchValve3,                             AcubeSATParameters::suControlSandwitchValve3},
-                  {AcubeSATParameters::SUFlowValveInlet1,                                    AcubeSATParameters::suFlowValveInlet1},
-                  {AcubeSATParameters::SUFlowValveInlet2,                                    AcubeSATParameters::suFlowValveInlet2},
-                  {AcubeSATParameters::SUFlowValveInlet3,                                    AcubeSATParameters::suFlowValveInlet3},
-                  {AcubeSATParameters::SUFlowValveOutlet1,                                   AcubeSATParameters::suFlowValveOutlet1},
-                  {AcubeSATParameters::SUFlowValveOutlet2,                                   AcubeSATParameters::suFlowValveOutlet2},
-                  {AcubeSATParameters::SUFlowValveOutlet3,                                   AcubeSATParameters::suFlowValveOutlet3},
-                  {AcubeSATParameters::SUFlowValveTest1,                                     AcubeSATParameters::suFlowValveTest1},
-                  {AcubeSATParameters::SUFlowValveTest2,                                     AcubeSATParameters::suFlowValveTest2},
-                  {AcubeSATParameters::SULEDIntensity1,                                      AcubeSATParameters::suLEDIntensity1},
-                  {AcubeSATParameters::SULEDIntensity2,                                      AcubeSATParameters::suLEDIntensity2},
-                  {AcubeSATParameters::SULEDIntensity3,                                      AcubeSATParameters::suLEDIntensity3},
-                  {AcubeSATParameters::SULEDIntensity4,                                      AcubeSATParameters::suLEDIntensity4},
-                  {AcubeSATParameters::SUNumberOfPhotosTaken,                                AcubeSATParameters::suNumberOfPhotosTaken},
-                  {AcubeSATParameters::SUPhotoExposureTime,                                  AcubeSATParameters::suPhotoExposureTime},
-                  {AcubeSATParameters::SUPhotoInterval,                                      AcubeSATParameters::suPhotoInterval},
-                  {AcubeSATParameters::SUPhotoCroppedResolution,                             AcubeSATParameters::suPhotoCroppedResolution},
-                  {AcubeSATParameters::SUPhotoCompressionRate,                               AcubeSATParameters::suPhotoCompressionRate},
-                  {AcubeSATParameters::SUExperimentNumber,                                   AcubeSATParameters::suExperimentNumber},
-                  {AcubeSATParameters::SUOnBoardTime,                                        AcubeSATParameters::suOnBoardTime},
-                  {AcubeSATParameters::SUTemperatureFromHumiditySensor1,                     AcubeSATParameters::suTemperatureFromHumiditySensor1},
-                  {AcubeSATParameters::SUTemperatureFromHumiditySensor2,                     AcubeSATParameters::suTemperatureFromHumiditySensor2},
+                  {PeakSatParameters::OBCPCBTemperature1,                                   PeakSatParameters::obcPCBTemperature1},
+                  {PeakSatParameters::OBCPCBTemperature2,                                   PeakSatParameters::obcPCBTemperature2},
+                  {PeakSatParameters::OBCMCUInputVoltage,                                   PeakSatParameters::obcMCUInputVoltage},
+                  {PeakSatParameters::OBCMCUBootCounter,                                    PeakSatParameters::obcMCUBootCounter},
+                  {PeakSatParameters::OBCFlashInt,                                          PeakSatParameters::obcFlashInt},
+                  {PeakSatParameters::OBCSRAMInt,                                           PeakSatParameters::obcSRAMInt},
+                  {PeakSatParameters::OBCAvailableMRAM,                                     PeakSatParameters::obcAvailableMRAM},
+                  {PeakSatParameters::OBCAvailableNAND,                                     PeakSatParameters::obcAvailableNAND},
+                  {PeakSatParameters::OBCSpacecraftTimeRef,                                 PeakSatParameters::obcSpacecraftTimeRef},
+                  {PeakSatParameters::OBCOnBoardTime,                                       PeakSatParameters::obcOnBoardTime},
+                  {PeakSatParameters::OBCOperationalMode,                                   PeakSatParameters::obcOperationalMode},
+                  {PeakSatParameters::OBCMemoryPartition,                                   PeakSatParameters::obcMemoryPartition},
+                  {PeakSatParameters::OBCReconfigurationTimer,                              PeakSatParameters::obcReconfigurationTimer},
+                  {PeakSatParameters::OBCLastFailedEvent,                                   PeakSatParameters::obcLastFailedEvent},
+                  {PeakSatParameters::OBCMCUSystick,                                        PeakSatParameters::obcMCUSystick},
+                  {PeakSatParameters::OBCCANBUSLoad1,                                       PeakSatParameters::obcCANBUSLoad1},
+                  {PeakSatParameters::OBCCANBUSLoad2,                                       PeakSatParameters::obcCANBUSLoad2},
+                  {PeakSatParameters::OBCCANBUSActive,                                      PeakSatParameters::obcCANBUSActive},
+                  {PeakSatParameters::OBCMCUFDIR,                                           PeakSatParameters::obcMCUFDIR},
+                  {PeakSatParameters::OBCMCURestartSafeModeThreshold,                       PeakSatParameters::obcMCURestartSafeModeThreshold},
+                  {PeakSatParameters::OBCNANDFLASHLCLThreshold,                             PeakSatParameters::obcNANDFLASHLCLThreshold},
+                  {PeakSatParameters::OBCMRAMLCLThreshold,                                  PeakSatParameters::obcMRAMLCLThreshold},
+                  {PeakSatParameters::OBCNANDFLASHON,                                       PeakSatParameters::obcNANDFLASHON},
+                  {PeakSatParameters::OBCMRAMON,                                            PeakSatParameters::obcNANDFlashScrubbingFrequency},
+                  {PeakSatParameters::OBCNANDFLASHScrubbingFrequency,                       PeakSatParameters::obcMRAMON},
+                  {PeakSatParameters::OBCRAMScrubbingFrequency,                             PeakSatParameters::obcRAMScrubbingFrequency},
+                  {PeakSatParameters::OBCProgramFlashScrubbingFrequency,                    PeakSatParameters::obcProgramFlashScrubbingFrequency},
+
             //---------------------------------------------------------------------------------------------
-                  {AcubeSATParameters::ADCSUseRTT,                                           AcubeSATParameters::adcsUseRTT},
-                  {AcubeSATParameters::ADCSUseUART,                                          AcubeSATParameters::adcsUseUART},
-                  {AcubeSATParameters::ADCSUseCAN,                                           AcubeSATParameters::adcsUseCAN},
-                  {AcubeSATParameters::ADCSXBodyFrameRegardingOrbitFrame,                    AcubeSATParameters::adcsXBodyFrameRegardingOrbitFrame},
-                  {AcubeSATParameters::ADCSYBodyFrameRegardingOrbitFrame,                    AcubeSATParameters::adcsYBodyFrameRegardingOrbitFrame},
-                  {AcubeSATParameters::ADCSZBodyFrameRegardingOrbitFrame,                    AcubeSATParameters::adcsZBodyFrameRegardingOrbitFrame},
-                  {AcubeSATParameters::ADCSAngularVelocityOnX,                               AcubeSATParameters::adcsAngularVelocityOnX},
-                  {AcubeSATParameters::ADCSAngularVelocityOnY,                               AcubeSATParameters::adcsAngularVelocityOnY},
-                  {AcubeSATParameters::ADCSAngularVelocityOnZ,                               AcubeSATParameters::adcsAngularVelocityOnZ},
-                  {AcubeSATParameters::ADCSBDotXAxis,                                        AcubeSATParameters::adcsBDotXAxis},
-                  {AcubeSATParameters::ADCSBDotYAxis,                                        AcubeSATParameters::adcsBDotYAxis},
-                  {AcubeSATParameters::ADCSBDotZAxis,                                        AcubeSATParameters::adcsBDotZAxis},
-                  {AcubeSATParameters::ADCSQuaternionOrbitFrameToBodyScalar,                 AcubeSATParameters::adcsQuaternionOrbitFrameToBodyScalar},
-                  {AcubeSATParameters::ADCSQuaternionOrbitFrameToBodyVectorElem1,            AcubeSATParameters::adcsQuaternionOrbitFrameToBodyVectorElem1},
-                  {AcubeSATParameters::ADCSQuaternionOrbitFrameToBodyVectorElem2,            AcubeSATParameters::adcsQuaternionOrbitFrameToBodyVectorElem2},
-                  {AcubeSATParameters::ADCSQuaternionOrbitFrameToBodyVectorElem3,            AcubeSATParameters::adcsQuaternionOrbitFrameToBodyVectorElem3},
-                  {AcubeSATParameters::ADCSMCUTemperature,                                   AcubeSATParameters::adcsMCUTemperature},
-                  {AcubeSATParameters::ADCSBootCounter,                                      AcubeSATParameters::adcsBootCounter},
-                  {AcubeSATParameters::ADCSOnBoardTime,                                      AcubeSATParameters::adcsOnBoardTime},
-                  {AcubeSATParameters::ADCSSystick,                                          AcubeSATParameters::adcsSystick},
-                  {AcubeSATParameters::ADCSMagnetometerRawX,                                 AcubeSATParameters::adcsMagnetometerRawX},
-                  {AcubeSATParameters::ADCSMagnetometerRawY,                                 AcubeSATParameters::adcsMagnetometerRawY},
-                  {AcubeSATParameters::ADCSMagnetometerRawZ,                                 AcubeSATParameters::adcsMagnetometerRawZ},
-                  {AcubeSATParameters::ADCSMagnetometerFrequency,                            AcubeSATParameters::adcsMagnetometerFrequency},
-                  {AcubeSATParameters::ADCSMagnetometerCycleCountX,                          AcubeSATParameters::adcsMagnetometerCycleCountX},
-                  {AcubeSATParameters::ADCSMagnetometerCycleCountY,                          AcubeSATParameters::adcsMagnetometerCycleCountY},
-                  {AcubeSATParameters::ADCSMagnetometerCycleCountZ,                          AcubeSATParameters::adcsMagnetometerCycleCountZ},
-                  {AcubeSATParameters::ADCSMagnetometerSelfTest,                             AcubeSATParameters::adcsMagnetometerSelfTest},
-                  {AcubeSATParameters::ADCSMagnetometerISISCalibratedXAxis,                  AcubeSATParameters::adcsMagnetometerISISCalibratedXAxis},
-                  {AcubeSATParameters::ADCSMagnetometerISISCalibratedYAxis,                  AcubeSATParameters::adcsMagnetometerISISCalibratedYAxis},
-                  {AcubeSATParameters::ADCSMagnetometerISISCalibratedZAxis,                  AcubeSATParameters::adcsMagnetometerISISCalibratedZAxis},
-                  {AcubeSATParameters::ADCSMagnetometerISISFilteredXAxis,                    AcubeSATParameters::adcsMagnetometerISISFilteredXAxis},
-                  {AcubeSATParameters::ADCSMagnetometerISISFilteredYAxis,                    AcubeSATParameters::adcsMagnetometerISISFilteredYAxis},
-                  {AcubeSATParameters::ADCSMagnetometerISISFilteredZAxis,                    AcubeSATParameters::adcsMagnetometerISISFilteredZAxis},
-                  {AcubeSATParameters::ADCSMagnetometerISISRawXAxis,                         AcubeSATParameters::adcsMagnetometerISISRawXAxis},
-                  {AcubeSATParameters::ADCSMagnetometerISISRawYAxis,                         AcubeSATParameters::adcsMagnetometerISISRawYAxis},
-                  {AcubeSATParameters::ADCSMagnetometerISISRawZAxis,                         AcubeSATParameters::adcsMagnetometerISISRawZAxis},
-                  {AcubeSATParameters::ADCSFineSunSensorRawTopLeft,                          AcubeSATParameters::adcsFineSunSensorRawTopLeft},
-                  {AcubeSATParameters::ADCSFineSunSensorRawTopRight,                         AcubeSATParameters::adcsFineSunSensorRawTopRight},
-                  {AcubeSATParameters::ADCSFineSunSensorRawBottomLeft,                       AcubeSATParameters::adcsFineSunSensorRawBottomLeft},
-                  {AcubeSATParameters::ADCSFineSunSensorRawBottomRight,                      AcubeSATParameters::adcsFineSunSensorRawBottomRight},
-                  {AcubeSATParameters::ADCSFineSunSensorFilteredTopLeft,                     AcubeSATParameters::adcsFineSunSensorFilteredTopLeft},
-                  {AcubeSATParameters::ADCSFineSunSensorFilteredTopRight,                    AcubeSATParameters::adcsFineSunSensorFilteredTopRight},
-                  {AcubeSATParameters::ADCSFineSunSensorFilteredBottomLeft,                  AcubeSATParameters::adcsFineSunSensorFilteredBottomLeft},
-                  {AcubeSATParameters::ADCSFineSunSensorFilteredBottomRight,                 AcubeSATParameters::adcsFineSunSensorFilteredBottomRight},
-                  {AcubeSATParameters::ADCSFineSunSensorAngularAlphaAngle,                   AcubeSATParameters::adcsFineSunSensorAngularAlphaAngle},
-                  {AcubeSATParameters::ADCSFineSunSensorAngularBetaAngle,                    AcubeSATParameters::adcsFineSunSensorAngularBetaAngle},
-                  {AcubeSATParameters::ADCSFineSunSensorCalculatedX,                         AcubeSATParameters::adcsFineSunSensorCalculatedX},
-                  {AcubeSATParameters::ADCSFineSunSensorCalculatedY,                         AcubeSATParameters::adcsFineSunSensorCalculatedY},
-                  {AcubeSATParameters::ADCSFineSunSensorCalculatedZ,                         AcubeSATParameters::adcsFineSunSensorCalculatedZ},
-                  {AcubeSATParameters::ADCSGyroscopeRateX,                                   AcubeSATParameters::adcsGyroscopeRateX},
-                  {AcubeSATParameters::ADCSGyroscopeRateY,                                   AcubeSATParameters::adcsGyroscopeRateY},
-                  {AcubeSATParameters::ADCSGyroscopeRateZ,                                   AcubeSATParameters::adcsGyroscopeRateZ},
-                  {AcubeSATParameters::ADCSGyroXLOCSTRegister,                               AcubeSATParameters::adcsGyroXLOCSTRegister},
-                  {AcubeSATParameters::ADCSGyroXHICSTRegister,                              AcubeSATParameters::adcsGyroXHICSTRegister},
-                  {AcubeSATParameters::ADCSGyroXQUADRegister,                               AcubeSATParameters::adcsGyroXQUADRegister},
-                  {AcubeSATParameters::ADCSGyroXFAULTRegister,                              AcubeSATParameters::adcsGyroXFAULTRegister},
-                  {AcubeSATParameters::ADCSGyroYLOCSTRegister,                              AcubeSATParameters::adcsGyroYLOCSTRegister},
-                  {AcubeSATParameters::ADCSGyroYHICSTRegister,                              AcubeSATParameters::adcsGyroYHICSTRegister},
-                  {AcubeSATParameters::ADCSGyroYQUADRegister,                               AcubeSATParameters::adcsGyroYQUADRegister},
-                  {AcubeSATParameters::ADCSGyroYFAULTRegister,                              AcubeSATParameters::adcsGyroYFAULTRegister},
-                  {AcubeSATParameters::ADCSGyroZLOCSTRegister,                              AcubeSATParameters::adcsGyroZLOCSTRegister},
-                  {AcubeSATParameters::ADCSGyroZHICSTRegister,                              AcubeSATParameters::adcsGyroZHICSTRegister},
-                  {AcubeSATParameters::ADCSGyroZQUADRegister,                               AcubeSATParameters::adcsGyroZQUADRegister},
-                  {AcubeSATParameters::ADCSGyroZFAULTRegister,                              AcubeSATParameters::adcsGyroZFAULTRegister},
-                  {AcubeSATParameters::ADCSMagnetorquerOnXAxisHBridgePolarity,               AcubeSATParameters::adcsMagnetorquerOnXAxisHBridgePolarity},
-                  {AcubeSATParameters::ADCSMagnetorquerOnYAxisHBridgePolarity,               AcubeSATParameters::adcsMagnetorquerOnYAxisHBridgePolarity},
-                  {AcubeSATParameters::ADCSMagnetorquerOnZAxisHBridgePolarity,               AcubeSATParameters::adcsMagnetorquerOnZAxisHBridgePolarity},
-                  {AcubeSATParameters::ADCSMagnetorquerOnXAxisOnOff,                         AcubeSATParameters::adcsMagnetorquerOnXAxisOnOff},
-                  {AcubeSATParameters::ADCSMagnetorquerOnYAxisOnOff,                         AcubeSATParameters::adcsMagnetorquerOnYAxisOnOff},
-                  {AcubeSATParameters::ADCSMagnetorquerOnZAxisOnOff,                         AcubeSATParameters::adcsMagnetorquerOnZAxisOnOff},
-                  {AcubeSATParameters::ADCSGainBdotXAxis,                                    AcubeSATParameters::adcsGainBdotXAxis},
-                  {AcubeSATParameters::ADCSGainBdotYAxis,                                    AcubeSATParameters::adcsGainBdotYAxis},
-                  {AcubeSATParameters::ADCSGainBdotZAxis,                                    AcubeSATParameters::adcsGainBdotZAxis},
-                  {AcubeSATParameters::ADCSGainProportionalPDXAxis,                          AcubeSATParameters::adcsGainProportionalPDXAxis},
-                  {AcubeSATParameters::ADCSGainProportionalPDYAxis,                          AcubeSATParameters::adcsGainProportionalPDYAxis},
-                  {AcubeSATParameters::ADCSGainProportionalPDZAxis,                          AcubeSATParameters::adcsGainProportionalPDZAxis},
-                  {AcubeSATParameters::ADCSGainDerivativePDXAxis,                            AcubeSATParameters::adcsGainDerivativePDXAxis},
-                  {AcubeSATParameters::ADCSGainDerivativePDYAxis,                            AcubeSATParameters::adcsGainDerivativePDYAxis},
-                  {AcubeSATParameters::ADCSGainDerivativePDZAxis,                            AcubeSATParameters::adcsGainDerivativePDZAxis},
-                  {AcubeSATParameters::ADCSiMQTBoardCoilXTemperatureSensor,                  AcubeSATParameters::adcsiMQTBoardCoilXTemperatureSensor},
-                  {AcubeSATParameters::ADCSiMQTBoardCoilYTemperatureSensor,                  AcubeSATParameters::adcsiMQTBoardCoilYTemperatureSensor},
-                  {AcubeSATParameters::ADCSiMQTBoardCoilZTemperatureSensor,                  AcubeSATParameters::adcsiMQTBoardCoilZTemperatureSensor},
-                  {AcubeSATParameters::ADCSiMQTBoardExtraTemperatureSensor,                  AcubeSATParameters::adcsiMQTBoardExtraTemperatureSensor},
-                  {AcubeSATParameters::ADCSGyroscopeXTemperature,                            AcubeSATParameters::adcsGyroscopeXTemperature},
-                  {AcubeSATParameters::ADCSGyroscopeYTemperature,                            AcubeSATParameters::adcsGyroscopeYTemperature},
-                  {AcubeSATParameters::ADCSGyroscopeZTemperature,                            AcubeSATParameters::adcsGyroscopeZTemperature},
-                  {AcubeSATParameters::ADCSBoardTemperature1,                                AcubeSATParameters::adcsBoardTemperature1},
-                  {AcubeSATParameters::ADCSBoardTemperature2,                                AcubeSATParameters::adcsBoardTemperature2},
-                  {AcubeSATParameters::ADCSMode,                                             AcubeSATParameters::adcsMode},
-                  {AcubeSATParameters::ADCSTLE,                                              AcubeSATParameters::adcsTLE},
-                  {AcubeSATParameters::ADCSInitialKalmanQuaternionECIFrameToBodyScalar,      AcubeSATParameters::adcsInitialKalmanQuaternionECIFrameToBodyScalar},
-                  {AcubeSATParameters::ADCSInitialKalmanQuaternionECIFrameToBodyVectorElem1, AcubeSATParameters::adcsInitialKalmanQuaternionECIFrameToBodyVectorElem1},
-                  {AcubeSATParameters::ADCSInitialKalmanQuaternionECIFrameToBodyVectorElem2, AcubeSATParameters::adcsInitialKalmanQuaternionECIFrameToBodyVectorElem2},
-                  {AcubeSATParameters::ADCSInitialKalmanQuaternionECIFrameToBodyVectorElem3, AcubeSATParameters::adcsInitialKalmanQuaternionECIFrameToBodyVectorElem3},
-                  {AcubeSATParameters::ADCSTLEOrbitalElementsArgumentPerigee,                AcubeSATParameters::adcsTLEOrbitalElementsArgumentPerigee},
-                  {AcubeSATParameters::ADCSTLEOrbitalElementsRAAN,                           AcubeSATParameters::adcsTLEOrbitalElementsRAAN},
-                  {AcubeSATParameters::ADCSTLEOrbitalElementsInclination,                    AcubeSATParameters::adcsTLEOrbitalElementsInclination},
-                  {AcubeSATParameters::ADCSTLEOrbitalElementsEccentricity,                   AcubeSATParameters::adcsTLEOrbitalElementsEccentricity},
-                  {AcubeSATParameters::ADCSTLEOrbitalElementsMeanAnomaly,                    AcubeSATParameters::adcsTLEOrbitalElementsMeanAnomaly},
-                  {AcubeSATParameters::ADCSSGP4OrbitalElementsArgumentPerigee,               AcubeSATParameters::adcsSGP4OrbitalElementsArgumentPerigee},
-                  {AcubeSATParameters::ADCSSGP4OrbitalElementsRAAN,                          AcubeSATParameters::adcsSGP4OrbitalElementsRAAN},
-                  {AcubeSATParameters::ADCSSGP4OrbitalElementsInclination,                   AcubeSATParameters::adcsSGP4OrbitalElementsInclination},
-                  {AcubeSATParameters::ADCSSGP4OrbitalElementsMeanAnomaly,                   AcubeSATParameters::adcsSGP4OrbitalElementsMeanAnomaly},
-                  {AcubeSATParameters::ADCSTimeGST_JD,                                       AcubeSATParameters::adcsTimeGST_JD},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesQ1,                   AcubeSATParameters::adcsDeterminationInitialMatricesQ1},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesQ2,                   AcubeSATParameters::adcsDeterminationInitialMatricesQ2},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesQ3,                   AcubeSATParameters::adcsDeterminationInitialMatricesQ3},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesQ4,                   AcubeSATParameters::adcsDeterminationInitialMatricesQ4},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesQ5,                   AcubeSATParameters::adcsDeterminationInitialMatricesQ5},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesQ6,                   AcubeSATParameters::adcsDeterminationInitialMatricesQ6},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesQ7,                   AcubeSATParameters::adcsDeterminationInitialMatricesQ7},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesQ8,                   AcubeSATParameters::adcsDeterminationInitialMatricesQ8},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesQ9,                   AcubeSATParameters::adcsDeterminationInitialMatricesQ9},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesQ10,                  AcubeSATParameters::adcsDeterminationInitialMatricesQ10},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesR1,                   AcubeSATParameters::adcsDeterminationInitialMatricesR1},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesR2,                   AcubeSATParameters::adcsDeterminationInitialMatricesR2},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesR3,                   AcubeSATParameters::adcsDeterminationInitialMatricesR3},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesR4,                   AcubeSATParameters::adcsDeterminationInitialMatricesR4},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesR5,                   AcubeSATParameters::adcsDeterminationInitialMatricesR5},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesR6,                   AcubeSATParameters::adcsDeterminationInitialMatricesR6},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesR7,                   AcubeSATParameters::adcsDeterminationInitialMatricesR7},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesR8,                   AcubeSATParameters::adcsDeterminationInitialMatricesR8},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesR9,                   AcubeSATParameters::adcsDeterminationInitialMatricesR9},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesPInitial1,            AcubeSATParameters::adcsDeterminationInitialMatricesPInitial1},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesPInitial2,            AcubeSATParameters::adcsDeterminationInitialMatricesPInitial2},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesPInitial3,            AcubeSATParameters::adcsDeterminationInitialMatricesPInitial3},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesPInitial4,            AcubeSATParameters::adcsDeterminationInitialMatricesPInitial4},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesPInitial5,            AcubeSATParameters::adcsDeterminationInitialMatricesPInitial5},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesPInitial6,            AcubeSATParameters::adcsDeterminationInitialMatricesPInitial6},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesPInitial7,            AcubeSATParameters::adcsDeterminationInitialMatricesPInitial7},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesPInitial8,            AcubeSATParameters::adcsDeterminationInitialMatricesPInitial8},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesPInitial9,            AcubeSATParameters::adcsDeterminationInitialMatricesPInitial9},
-                  {AcubeSATParameters::ADCSDeterminationInitialMatricesPInitial10,           AcubeSATParameters::adcsDeterminationInitialMatricesPInitial10},
-                  {AcubeSATParameters::ADCSReactionWheelOn,                                  AcubeSATParameters::adcsReactionWheelOn},
-                  {AcubeSATParameters::ADCSReactionWheelAngularVelocity,                     AcubeSATParameters::adcsReactionWheelAngularVelocity},
-                  {AcubeSATParameters::ADCSMagnetorquerAxisAssignment,                       AcubeSATParameters::adcsMagnetorquerAxisAssignment},
-                  {AcubeSATParameters::ADCSControlProfileSwitching,                          AcubeSATParameters::adcsControlProfileSwitching},
-                  {AcubeSATParameters::ADCSTransformationAxisXAssignment,                    AcubeSATParameters::adcsTransformationAxisXAssignment},
-                  {AcubeSATParameters::ADCSTransformationAxisYAssignment,                    AcubeSATParameters::adcsTransformationAxisYAssignment},
-                  {AcubeSATParameters::ADCSTransformationAxisZAssignment,                    AcubeSATParameters::adcsTransformationAxisZAssignment},
-                  {AcubeSATParameters::ADCSMagnetometerAxisAssignment,                       AcubeSATParameters::adcsMagnetometerAxisAssignment},
-                  {AcubeSATParameters::ADCSSunSensorAxisAssignment,                          AcubeSATParameters::adcsSunSensorAxisAssignment},
-                  {AcubeSATParameters::ADCSGyroscopeAxisAssignment,                          AcubeSATParameters::adcsGyroscopeAxisAssignment},
-                  {AcubeSATParameters::ADCSFrameAxisSignECIToECEF_X,                         AcubeSATParameters::adcsFrameAxisSignECIToECEF_X},
-                  {AcubeSATParameters::ADCSFrameAxisSignECIToECEF_Y,                         AcubeSATParameters::adcsFrameAxisSignECIToECEF_Y},
-                  {AcubeSATParameters::ADCSFrameAxisSignECIToECEF_Z,                         AcubeSATParameters::adcsFrameAxisSignECIToECEF_Z},
-                  {AcubeSATParameters::ADCSFrameAxisSignECIToOrbit_X,                        AcubeSATParameters::adcsFrameAxisSignECIToOrbit_X},
-                  {AcubeSATParameters::ADCSFrameAxisSignECIToOrbit_Y,                        AcubeSATParameters::adcsFrameAxisSignECIToOrbit_Y},
-                  {AcubeSATParameters::ADCSFrameAxisSignECIToOrbit_Z,                        AcubeSATParameters::adcsFrameAxisSignECIToOrbit_Z},
-                  {AcubeSATParameters::ADCSFrameAxisSignNEDToECEF_X,                         AcubeSATParameters::adcsFrameAxisSignNEDToECEF_X},
-                  {AcubeSATParameters::ADCSFrameAxisSignNEDToECEF_Y,                         AcubeSATParameters::adcsFrameAxisSignNEDToECEF_Y},
-                  {AcubeSATParameters::ADCSFrameAxisSignNEDToECEF_Z,                         AcubeSATParameters::adcsFrameAxisSignNEDToECEF_Z},
-                  {AcubeSATParameters::ADCSMagnetometerSignX,                                AcubeSATParameters::adcsMagnetometerSignX},
-                  {AcubeSATParameters::ADCSMagnetometerSignY,                                AcubeSATParameters::adcsMagnetometerSignY},
-                  {AcubeSATParameters::ADCSMagnetometerSignZ,                                AcubeSATParameters::adcsMagnetometerSignZ},
-                  {AcubeSATParameters::ADCSSunSensorXAxisSign,                               AcubeSATParameters::adcsSunSensorXAxisSign},
-                  {AcubeSATParameters::ADCSSunSensorYAxisSign,                               AcubeSATParameters::adcsSunSensorYAxisSign},
-                  {AcubeSATParameters::ADCSSunSensorZAxisSign,                               AcubeSATParameters::adcsSunSensorZAxisSign},
-                  {AcubeSATParameters::ADCSGyroSignX,                                        AcubeSATParameters::adcsGyroSignX},
-                  {AcubeSATParameters::ADCSGyroSignY,                                        AcubeSATParameters::adcsGyroSignY},
-                  {AcubeSATParameters::ADCSGyroSignZ,                                        AcubeSATParameters::adcsGyroSignZ},
-                  {AcubeSATParameters::ADCSDetumblingNominalThresholds,                      AcubeSATParameters::adcsDetumblingNominalThresholds},
-                  {AcubeSATParameters::ADCSNominalDetumblingThresholds,                      AcubeSATParameters::adcsNominalDetumblingThresholds},
-                  {AcubeSATParameters::ADCSStandByDetumblingThresholds,                      AcubeSATParameters::adcsStandByDetumblingThresholds},
-                  {AcubeSATParameters::ADCSDetumblingStandByThresholds,                      AcubeSATParameters::adcsDetumblingStandByThresholds},
-                  {AcubeSATParameters::ADCSNominalStandByThresholds,                         AcubeSATParameters::adcsNominalStandByThresholds},
-                  {AcubeSATParameters::ADCSMagnetorquerMagneticDipoleX,                      AcubeSATParameters::adcsMagnetorquerMagneticDipoleX},
-                  {AcubeSATParameters::ADCSMagnetorquerMagneticDipoleY,                      AcubeSATParameters::adcsMagnetorquerMagneticDipoleY},
-                  {AcubeSATParameters::ADCSMagnetorquerMagneticDipoleZ,                      AcubeSATParameters::adcsMagnetorquerMagneticDipoleZ},
-                  {AcubeSATParameters::ADCSMagnetorquerDutyCycle,                            AcubeSATParameters::adcsMagnetorquerDutyCycle},
-                  {AcubeSATParameters::ADCSMagnetorquerCurrentX,                             AcubeSATParameters::adcsMagnetorquerCurrentX},
-                  {AcubeSATParameters::ADCSMagnetorquerCurrentY,                             AcubeSATParameters::adcsMagnetorquerCurrentY},
-                  {AcubeSATParameters::ADCSMagnetorquerCurrentZ,                             AcubeSATParameters::adcsMagnetorquerCurrentZ},
-                  {AcubeSATParameters::ADCSMagnetorquerBDot,                                 AcubeSATParameters::adcsMagnetorquerBDot},
-                  {AcubeSATParameters::ADCSReactionWheelTorque,                              AcubeSATParameters::adcsReactionWheelTorque},
-                  {AcubeSATParameters::ADCSReactionWheelAngularAcceleration,                 AcubeSATParameters::adcsReactionWheelAngularAcceleration},
-                  {AcubeSATParameters::ADCSDesiredQuaternionOrbitToBodyScalar,               AcubeSATParameters::adcsDesiredQuaternionOrbitToBodyScalar},
-                  {AcubeSATParameters::ADCSDesiredQuaternionVectorElem1,                     AcubeSATParameters::adcsDesiredQuaternionVectorElem1},
-                  {AcubeSATParameters::ADCSDesiredQuaternionVectorElem2,                     AcubeSATParameters::adcsDesiredQuaternionVectorElem2},
-                  {AcubeSATParameters::ADCSDesiredQuaternionVectorElem3,                     AcubeSATParameters::adcsDesiredQuaternionVectorElem3},
-                  {AcubeSATParameters::ADCSAngularVelocityDesiredX,                          AcubeSATParameters::adcsAngularVelocityDesiredX},
-                  {AcubeSATParameters::ADCSAngularVelocityDesiredY,                          AcubeSATParameters::adcsAngularVelocityDesiredY},
-                  {AcubeSATParameters::ADCSAngularVelocityDesiredZ,                          AcubeSATParameters::adcsAngularVelocityDesiredZ},
-                  {AcubeSATParameters::ADCSGyroBiasX,                                        AcubeSATParameters::adcsGyroBiasX},
-                  {AcubeSATParameters::ADCSGyroBiasY,                                        AcubeSATParameters::adcsGyroBiasY},
-                  {AcubeSATParameters::ADCSGyroBiasZ,                                        AcubeSATParameters::adcsGyroBiasZ},
-                  {AcubeSATParameters::ADCSSatellitePositionLatitude,                        AcubeSATParameters::adcsSatellitePositionLatitude},
-                  {AcubeSATParameters::ADCSSatellitePositionLongitude,                       AcubeSATParameters::adcsSatellitePositionLongitude},
-                  {AcubeSATParameters::ADCSSatellitePositionHeight,                          AcubeSATParameters::adcsSatellitePositionHeight},
-                  {AcubeSATParameters::ADCSSatellitePositionECIX,                            AcubeSATParameters::adcsSatellitePositionECIX},
-                  {AcubeSATParameters::ADCSSatellitePositionECIY,                            AcubeSATParameters::adcsSatellitePositionECIY},
-                  {AcubeSATParameters::ADCSSatellitePositionECIZ,                            AcubeSATParameters::adcsSatellitePositionECIZ},
-                  {AcubeSATParameters::ADCSSatelliteLinearVelocityECIX,                      AcubeSATParameters::adcsSatelliteLinearVelocityECIX},
-                  {AcubeSATParameters::ADCSSatelliteLinearVelocityECIY,                      AcubeSATParameters::adcsSatelliteLinearVelocityECIY},
-                  {AcubeSATParameters::ADCSSatelliteLinearVelocityECIZ,                      AcubeSATParameters::adcsSatelliteLinearVelocityECIZ},
-                  {AcubeSATParameters::ADCSMagneticFieldNEDX,                                AcubeSATParameters::adcsMagneticFieldNEDX},
-                  {AcubeSATParameters::ADCSMagneticFieldNEDY,                                AcubeSATParameters::adcsMagneticFieldNEDY},
-                  {AcubeSATParameters::ADCSMagneticFieldNEDZ,                                AcubeSATParameters::adcsMagneticFieldNEDZ},
-                  {AcubeSATParameters::ADCSMagneticFieldECIX,                                AcubeSATParameters::adcsMagneticFieldECIX},
-                  {AcubeSATParameters::ADCSMagneticFieldECIY,                                AcubeSATParameters::adcsMagneticFieldECIY},
-                  {AcubeSATParameters::ADCSMagneticFieldECIZ,                                AcubeSATParameters::adcsMagneticFieldECIZ},
-                  {AcubeSATParameters::ADCSSunPositionECIX,                                  AcubeSATParameters::adcsSunPositionECIX},
-                  {AcubeSATParameters::ADCSSunPositionECIY,                                  AcubeSATParameters::adcsSunPositionECIY},
-                  {AcubeSATParameters::ADCSSunPositionECIZ,                                  AcubeSATParameters::adcsSunPositionECIZ},
-                  {AcubeSATParameters::ADCSEclipse,                                          AcubeSATParameters::adcsEclipse},
-                  {AcubeSATParameters::ADCSDesaturationPecentageParameter,                   AcubeSATParameters::adcsDesaturationPecentageParameter},
-                  {AcubeSATParameters::ADCSISISMTQModes,                                     AcubeSATParameters::adcsISISMTQModes},
-                  {AcubeSATParameters::ADCSMagnetorquerActuationMethod,                      AcubeSATParameters::adcsMagnetorquerActuationMethod},
-                  {AcubeSATParameters::ADCSDutyCycleDetermination,                           AcubeSATParameters::adcsDutyCycleDetermination},
-                  {AcubeSATParameters::ADCSDutyCycleControl,                                 AcubeSATParameters::adcsDutyCycleControl},
-                  {AcubeSATParameters::ADCSDutyCycleOther,                                   AcubeSATParameters::adcsDutyCycleOther},
-                  {AcubeSATParameters::ADCSDesatInitialAngularVelocityParam,                 AcubeSATParameters::adcsDesatInitialAngularVelocityParam},
-                  {AcubeSATParameters::ADCSMCUInputVoltage,                                  AcubeSATParameters::adcsMCUInputVoltage},
-                  {AcubeSATParameters::ADCSMemoryPartition,                                  AcubeSATParameters::adcsMemoryPartition},
-                  {AcubeSATParameters::ADCSLastFailedEvent,                                  AcubeSATParameters::adcsLastFailedEvent},
-                  {AcubeSATParameters::ADCSRAMScrubbingFrequency,                            AcubeSATParameters::adcsRAMScrubbingFrequency},
-                  {AcubeSATParameters::ADCSProgramFlashScrubbingFrequency,                   AcubeSATParameters::adcsProgramFlashScrubbingFrequency},
-                  {AcubeSATParameters::ADCSFLASHInt,                                         AcubeSATParameters::adcsFlashInt},
-                  {AcubeSATParameters::ADCSSRAMInt,                                          AcubeSATParameters::adcsSRAMInt},
+                  {PeakSatParameters::EPSSwitch3V1,                                         PeakSatParameters::epsSwitch3V1},
+                  {PeakSatParameters::EPSSwitch3V2,                                         PeakSatParameters::epsSwitch3V2},
+                  {PeakSatParameters::EPSSwitch3V3,                                         PeakSatParameters::epsSwitch3V3},
+                  {PeakSatParameters::EPSSwitch3V4,                                         PeakSatParameters::epsSwitch3V4},
+                  {PeakSatParameters::EPSSwitch5V1,                                         PeakSatParameters::epsSwitch5V1},
+                  {PeakSatParameters::EPSSwitch5V2,                                         PeakSatParameters::epsSwitch5V2},
+                  {PeakSatParameters::EPSSwitch5V3,                                         PeakSatParameters::epsSwitch5V3},
+                  {PeakSatParameters::EPSSwitch5V4,                                         PeakSatParameters::epsSwitch5V4},
+                  {PeakSatParameters::EPSSwitch1,                                           PeakSatParameters::epsSwitch1},
+                  {PeakSatParameters::EPSSwitch2,                                           PeakSatParameters::epsSwitch2},
+                  {PeakSatParameters::EPSSwitch3,                                           PeakSatParameters::epsSwitch3},
+                  {PeakSatParameters::EPSSwitch4,                                           PeakSatParameters::epsSwitch4},
+                  {PeakSatParameters::EPSSwitch5,                                           PeakSatParameters::epsSwitch5},
+                  {PeakSatParameters::EPSCurrentForRail3V1,                                 PeakSatParameters::epsCurrentForRail3V1},
+                  {PeakSatParameters::EPSCurrentForRail3V2,                                 PeakSatParameters::epsCurrentForRail3V2},
+                  {PeakSatParameters::EPSCurrentForRail3V3,                                 PeakSatParameters::epsCurrentForRail3V3},
+                  {PeakSatParameters::EPSCurrentForRail3V4,                                 PeakSatParameters::epsCurrentForRail3V4},
+                  {PeakSatParameters::EPSCurrentForRail5V1,                                 PeakSatParameters::epsCurrentForRail5V1},
+                  {PeakSatParameters::EPSCurrentForRail5V2,                                 PeakSatParameters::epsCurrentForRail5V2},
+                  {PeakSatParameters::EPSCurrentForRail5V3,                                 PeakSatParameters::epsCurrentForRail5V3},
+                  {PeakSatParameters::EPSCurrentForRail5V4,                                 PeakSatParameters::epsCurrentForRail5V4},
+                  {PeakSatParameters::EPSCurrentForRail1,                                   PeakSatParameters::epsCurrentForRail1},
+                  {PeakSatParameters::EPSCurrentForRail2,                                   PeakSatParameters::epsCurrentForRail2},
+                  {PeakSatParameters::EPSCurrentForRail3,                                   PeakSatParameters::epsCurrentForRail3},
+                  {PeakSatParameters::EPSCurrentForRail4,                                   PeakSatParameters::epsCurrentForRail4},
+                  {PeakSatParameters::EPSCurrentForRail5,                                   PeakSatParameters::epsCurrentForRail5},
+                  {PeakSatParameters::EPSVoltageForEachRail3V1,                             PeakSatParameters::epsVoltageForEachRail3V1},
+                  {PeakSatParameters::EPSVoltageForEachRail3V2,                             PeakSatParameters::epsVoltageForEachRail3V2},
+                  {PeakSatParameters::EPSVoltageForEachRail3V3,                             PeakSatParameters::epsVoltageForEachRail3V3},
+                  {PeakSatParameters::EPSVoltageForEachRail3V4,                             PeakSatParameters::epsVoltageForEachRail3V4},
+                  {PeakSatParameters::EPSVoltageForEachRail5V1,                             PeakSatParameters::epsVoltageForEachRail5V1},
+                  {PeakSatParameters::EPSVoltageForEachRail5V2,                             PeakSatParameters::epsVoltageForEachRail5V2},
+                  {PeakSatParameters::EPSVoltageForEachRail5V3,                             PeakSatParameters::epsVoltageForEachRail5V3},
+                  {PeakSatParameters::EPSVoltageForEachRail5V4,                             PeakSatParameters::epsVoltageForEachRail5V4},
+                  {PeakSatParameters::EPSVoltageForEachRail12V1,                            PeakSatParameters::epsVoltageForEachRail12V1},
+                  {PeakSatParameters::EPSVoltageForEachRail12V2,                            PeakSatParameters::epsVoltageForEachRail12V2},
+                  {PeakSatParameters::EPSVoltageForEachRail12V3,                            PeakSatParameters::epsVoltageForEachRail12V3},
+                  {PeakSatParameters::EPSVoltageForEachRail12V4,                            PeakSatParameters::epsVoltageForEachRail12V4},
+                  {PeakSatParameters::EPSPanelVoltageXPlus,                                 PeakSatParameters::epsPanelVoltageXPlus},
+                  {PeakSatParameters::EPSPanelVoltageXMinus,                                PeakSatParameters::epsPanelVoltageXMinus},
+                  {PeakSatParameters::EPSPanelVoltageYPlus,                                 PeakSatParameters::epsPanelVoltageYPlus},
+                  {PeakSatParameters::EPSPanelVoltageYMinus,                                PeakSatParameters::epsPanelVoltageYMinus},
+                  {PeakSatParameters::EPSPanelVoltageZ,                                     PeakSatParameters::epsPanelVoltageZ},
+                  {PeakSatParameters::EPSPanelCurrentXPlus,                                 PeakSatParameters::epsPanelCurrentXPlus},
+                  {PeakSatParameters::EPSPanelCurrentXMinus,                                PeakSatParameters::epsPanelCurrentXMinus},
+                  {PeakSatParameters::EPSPanelCurrentYPlus,                                 PeakSatParameters::epsPanelCurrentYPlus},
+                  {PeakSatParameters::EPSPanelCurrentYMinus,                                PeakSatParameters::epsPanelCurrentYMinus},
+                  {PeakSatParameters::EPSPanelCurrentZ,                                     PeakSatParameters::epsPanelCurrentZ},
+                  {PeakSatParameters::EPSBatteryPackVoltage,                                PeakSatParameters::epsBatteryPackVoltage},
+                  {PeakSatParameters::EPSPanelTemperatureXPlus,                             PeakSatParameters::epsPanelTemperatureXPlus},
+                  {PeakSatParameters::EPSPanelTemperatureXMinus,                            PeakSatParameters::epsPanelTemperatureXMinus},
+                  {PeakSatParameters::EPSPanelTemperatureYPlus,                             PeakSatParameters::epsPanelTemperatureYPlus},
+                  {PeakSatParameters::EPSPanelTemperatureYMinus,                            PeakSatParameters::epsPanelTemperatureYMinus},
+                  {PeakSatParameters::EPSPanelTemperatureZ,                                 PeakSatParameters::epsPanelTemperatureZ},
+                  {PeakSatParameters::EPSPCBTemperatureSensor1,                             PeakSatParameters::epsPCBTemperatureSensor1},
+                  {PeakSatParameters::EPSPCBTemperatureSensor2,                             PeakSatParameters::epsPCBTemperatureSensor2},
+                  {PeakSatParameters::EPSBatteryTemperatureSensor1,                         PeakSatParameters::epsBatteryTemperatureSensor1},
+                  {PeakSatParameters::EPSBatteryTemperatureSensor2,                         PeakSatParameters::epsBatteryTemperatureSensor2},
             //---------------------------------------------------------------------------------------------
-                  {AcubeSATParameters::EPSSwitch3V1,                                         AcubeSATParameters::epsSwitch3V1},
-                  {AcubeSATParameters::EPSSwitch3V2,                                         AcubeSATParameters::epsSwitch3V2},
-                  {AcubeSATParameters::EPSSwitch3V3,                                         AcubeSATParameters::epsSwitch3V3},
-                  {AcubeSATParameters::EPSSwitch3V4,                                         AcubeSATParameters::epsSwitch3V4},
-                  {AcubeSATParameters::EPSSwitch5V1,                                         AcubeSATParameters::epsSwitch5V1},
-                  {AcubeSATParameters::EPSSwitch5V2,                                         AcubeSATParameters::epsSwitch5V2},
-                  {AcubeSATParameters::EPSSwitch5V3,                                         AcubeSATParameters::epsSwitch5V3},
-                  {AcubeSATParameters::EPSSwitch5V4,                                         AcubeSATParameters::epsSwitch5V4},
-                  {AcubeSATParameters::EPSSwitch1,                                           AcubeSATParameters::epsSwitch1},
-                  {AcubeSATParameters::EPSSwitch2,                                           AcubeSATParameters::epsSwitch2},
-                  {AcubeSATParameters::EPSSwitch3,                                           AcubeSATParameters::epsSwitch3},
-                  {AcubeSATParameters::EPSSwitch4,                                           AcubeSATParameters::epsSwitch4},
-                  {AcubeSATParameters::EPSSwitch5,                                           AcubeSATParameters::epsSwitch5},
-                  {AcubeSATParameters::EPSCurrentForRail3V1,                                 AcubeSATParameters::epsCurrentForRail3V1},
-                  {AcubeSATParameters::EPSCurrentForRail3V2,                                 AcubeSATParameters::epsCurrentForRail3V2},
-                  {AcubeSATParameters::EPSCurrentForRail3V3,                                 AcubeSATParameters::epsCurrentForRail3V3},
-                  {AcubeSATParameters::EPSCurrentForRail3V4,                                 AcubeSATParameters::epsCurrentForRail3V4},
-                  {AcubeSATParameters::EPSCurrentForRail5V1,                                 AcubeSATParameters::epsCurrentForRail5V1},
-                  {AcubeSATParameters::EPSCurrentForRail5V2,                                 AcubeSATParameters::epsCurrentForRail5V2},
-                  {AcubeSATParameters::EPSCurrentForRail5V3,                                 AcubeSATParameters::epsCurrentForRail5V3},
-                  {AcubeSATParameters::EPSCurrentForRail5V4,                                 AcubeSATParameters::epsCurrentForRail5V4},
-                  {AcubeSATParameters::EPSCurrentForRail1,                                   AcubeSATParameters::epsCurrentForRail1},
-                  {AcubeSATParameters::EPSCurrentForRail2,                                   AcubeSATParameters::epsCurrentForRail2},
-                  {AcubeSATParameters::EPSCurrentForRail3,                                   AcubeSATParameters::epsCurrentForRail3},
-                  {AcubeSATParameters::EPSCurrentForRail4,                                   AcubeSATParameters::epsCurrentForRail4},
-                  {AcubeSATParameters::EPSCurrentForRail5,                                   AcubeSATParameters::epsCurrentForRail5},
-                  {AcubeSATParameters::EPSVoltageForEachRail3V1,                             AcubeSATParameters::epsVoltageForEachRail3V1},
-                  {AcubeSATParameters::EPSVoltageForEachRail3V2,                             AcubeSATParameters::epsVoltageForEachRail3V2},
-                  {AcubeSATParameters::EPSVoltageForEachRail3V3,                             AcubeSATParameters::epsVoltageForEachRail3V3},
-                  {AcubeSATParameters::EPSVoltageForEachRail3V4,                             AcubeSATParameters::epsVoltageForEachRail3V4},
-                  {AcubeSATParameters::EPSVoltageForEachRail5V1,                             AcubeSATParameters::epsVoltageForEachRail5V1},
-                  {AcubeSATParameters::EPSVoltageForEachRail5V2,                             AcubeSATParameters::epsVoltageForEachRail5V2},
-                  {AcubeSATParameters::EPSVoltageForEachRail5V3,                             AcubeSATParameters::epsVoltageForEachRail5V3},
-                  {AcubeSATParameters::EPSVoltageForEachRail5V4,                             AcubeSATParameters::epsVoltageForEachRail5V4},
-                  {AcubeSATParameters::EPSVoltageForEachRail12V1,                            AcubeSATParameters::epsVoltageForEachRail12V1},
-                  {AcubeSATParameters::EPSVoltageForEachRail12V2,                            AcubeSATParameters::epsVoltageForEachRail12V2},
-                  {AcubeSATParameters::EPSVoltageForEachRail12V3,                            AcubeSATParameters::epsVoltageForEachRail12V3},
-                  {AcubeSATParameters::EPSVoltageForEachRail12V4,                            AcubeSATParameters::epsVoltageForEachRail12V4},
-                  {AcubeSATParameters::EPSPanelVoltageXPlus,                                 AcubeSATParameters::epsPanelVoltageXPlus},
-                  {AcubeSATParameters::EPSPanelVoltageXMinus,                                AcubeSATParameters::epsPanelVoltageXMinus},
-                  {AcubeSATParameters::EPSPanelVoltageYPlus,                                 AcubeSATParameters::epsPanelVoltageYPlus},
-                  {AcubeSATParameters::EPSPanelVoltageYMinus,                                AcubeSATParameters::epsPanelVoltageYMinus},
-                  {AcubeSATParameters::EPSPanelVoltageZ,                                     AcubeSATParameters::epsPanelVoltageZ},
-                  {AcubeSATParameters::EPSPanelCurrentXPlus,                                 AcubeSATParameters::epsPanelCurrentXPlus},
-                  {AcubeSATParameters::EPSPanelCurrentXMinus,                                AcubeSATParameters::epsPanelCurrentXMinus},
-                  {AcubeSATParameters::EPSPanelCurrentYPlus,                                 AcubeSATParameters::epsPanelCurrentYPlus},
-                  {AcubeSATParameters::EPSPanelCurrentYMinus,                                AcubeSATParameters::epsPanelCurrentYMinus},
-                  {AcubeSATParameters::EPSPanelCurrentZ,                                     AcubeSATParameters::epsPanelCurrentZ},
-                  {AcubeSATParameters::EPSBatteryPackVoltage,                                AcubeSATParameters::epsBatteryPackVoltage},
-                  {AcubeSATParameters::EPSPanelTemperatureXPlus,                             AcubeSATParameters::epsPanelTemperatureXPlus},
-                  {AcubeSATParameters::EPSPanelTemperatureXMinus,                            AcubeSATParameters::epsPanelTemperatureXMinus},
-                  {AcubeSATParameters::EPSPanelTemperatureYPlus,                             AcubeSATParameters::epsPanelTemperatureYPlus},
-                  {AcubeSATParameters::EPSPanelTemperatureYMinus,                            AcubeSATParameters::epsPanelTemperatureYMinus},
-                  {AcubeSATParameters::EPSPanelTemperatureZ,                                 AcubeSATParameters::epsPanelTemperatureZ},
-                  {AcubeSATParameters::EPSPCBTemperatureSensor1,                             AcubeSATParameters::epsPCBTemperatureSensor1},
-                  {AcubeSATParameters::EPSPCBTemperatureSensor2,                             AcubeSATParameters::epsPCBTemperatureSensor2},
-                  {AcubeSATParameters::EPSBatteryTemperatureSensor1,                         AcubeSATParameters::epsBatteryTemperatureSensor1},
-                  {AcubeSATParameters::EPSBatteryTemperatureSensor2,                         AcubeSATParameters::epsBatteryTemperatureSensor2},
-            //---------------------------------------------------------------------------------------------
-                  {AcubeSATParameters::COMMSUHFBandPATemperature,                            AcubeSATParameters::commsUHFBandPATemperature},
-                  {AcubeSATParameters::COMMSSBandPATemperature,                              AcubeSATParameters::commsSBandPATemperature},
-                  {AcubeSATParameters::COMMSPCBTemperature,                                  AcubeSATParameters::commsPCBTemperature},
-                  {AcubeSATParameters::COMMSAntennaDeploymentStatus,                         AcubeSATParameters::commsAntennaDeploymentStatus},
-                  {AcubeSATParameters::COMMSDataRateUHFTX,                                   AcubeSATParameters::commsDataRateUHFTX},
-                  {AcubeSATParameters::COMMSDataRateUHFRX,                                   AcubeSATParameters::commsDataRateUHFRX},
-                  {AcubeSATParameters::COMMSSymbolRateSBand,                                 AcubeSATParameters::commsSymbolRateSBand},
-                  {AcubeSATParameters::COMMSCWInterval,                                      AcubeSATParameters::commsCWInterval},
-                  {AcubeSATParameters::COMMSGMSKBeaconInterval,                              AcubeSATParameters::commsGMSKBeaconInterval},
-                  {AcubeSATParameters::COMMSUHFBandTXPower,                                  AcubeSATParameters::commsUHFBandTXPower},
-                  {AcubeSATParameters::COMMSSBandTXPower,                                    AcubeSATParameters::commsSBandTXPower},
-                  {AcubeSATParameters::COMMSChannelNumberUHFBand,                            AcubeSATParameters::commsChannelNumberUHFBand},
-                  {AcubeSATParameters::COMMSChannelNumberSBand,                              AcubeSATParameters::commsChannelNumberSBand},
-                  {AcubeSATParameters::COMMSLNAGain,                                         AcubeSATParameters::commsLNAGain},
-                  {AcubeSATParameters::COMMSPAGainUHFBand,                                   AcubeSATParameters::commsPAGainUHFBand},
-                  {AcubeSATParameters::COMMSPAGainSBand,                                     AcubeSATParameters::commsPAGainSBand},
-                  {AcubeSATParameters::COMMSVGAGain,                                         AcubeSATParameters::commsVGAGain},
-                  {AcubeSATParameters::COMMSRSSI,                                            AcubeSATParameters::commsRSSI},
-                  {AcubeSATParameters::COMMSUHFBandTXOnOff,                                  AcubeSATParameters::commsUHFBandTXOnOff},
-                  {AcubeSATParameters::COMMSUHFBandRXOnOff,                                  AcubeSATParameters::commsUHFBandRXOnOff},
-                  {AcubeSATParameters::COMMSSBandTXOnOff,                                    AcubeSATParameters::commsSBandTXOnOff},
-                  {AcubeSATParameters::COMMSPacketsRejectedCOMMS,                            AcubeSATParameters::commsPacketsRejectedCOMMS},
-                  {AcubeSATParameters::COMMSInvalidHMAC,                                     AcubeSATParameters::commsInvalidHMAC},
-                  {AcubeSATParameters::COMMSInvalidPacketStructure,                          AcubeSATParameters::commsInvalidPacketStructure},
-                  {AcubeSATParameters::COMMSInvalidSpacecraftID,                             AcubeSATParameters::commsInvalidSpacecraftID},
-                  {AcubeSATParameters::COMMSFrameSequenceCounter,                            AcubeSATParameters::commsFrameSequenceCounter},
-                  {AcubeSATParameters::COMMSPCBTemperature1,                                 AcubeSATParameters::commsPCBTemperature1},
-                  {AcubeSATParameters::COMMSPCBTemperature2,                                 AcubeSATParameters::commsPCBTemperature2},
-                  {AcubeSATParameters::COMMSMCUTemperature,                                  AcubeSATParameters::commsMCUTemperature},
-                  {AcubeSATParameters::COMMSMCUInputVoltage,                                 AcubeSATParameters::commsMCUInputVoltage},
-                  {AcubeSATParameters::COMMSMCUBootCounter,                                  AcubeSATParameters::commsMCUBootCounter},
-                  {AcubeSATParameters::COMMSOnBoardTime,                                     AcubeSATParameters::commsOnBoardTime},
-                  {AcubeSATParameters::COMMSNANDCurrentlyUsedMemoryPartition,                AcubeSATParameters::commsNANDCurrentlyUsedMemoryPartition},
-                  {AcubeSATParameters::COMMSLastFailedEvent,                                 AcubeSATParameters::commsLastFailedEvent},
-                  {AcubeSATParameters::COMMSMCUSystick,                                      AcubeSATParameters::commsMCUSystick},
-                  {AcubeSATParameters::COMMSFlashInt,                                        AcubeSATParameters::commsFlashInt},
-                  {AcubeSATParameters::COMMSSRAMInt,                                         AcubeSATParameters::commsSRAMInt}};
+                  {PeakSatParameters::COMMSUHFBandPATemperature,                            PeakSatParameters::commsUHFBandPATemperature},
+                  {PeakSatParameters::COMMSSBandPATemperature,                              PeakSatParameters::commsSBandPATemperature},
+                  {PeakSatParameters::COMMSPCBTemperature,                                  PeakSatParameters::commsPCBTemperature},
+                  {PeakSatParameters::COMMSAntennaDeploymentStatus,                         PeakSatParameters::commsAntennaDeploymentStatus},
+                  {PeakSatParameters::COMMSDataRateUHFTX,                                   PeakSatParameters::commsDataRateUHFTX},
+                  {PeakSatParameters::COMMSDataRateUHFRX,                                   PeakSatParameters::commsDataRateUHFRX},
+                  {PeakSatParameters::COMMSSymbolRateSBand,                                 PeakSatParameters::commsSymbolRateSBand},
+                  {PeakSatParameters::COMMSCWInterval,                                      PeakSatParameters::commsCWInterval},
+                  {PeakSatParameters::COMMSGMSKBeaconInterval,                              PeakSatParameters::commsGMSKBeaconInterval},
+                  {PeakSatParameters::COMMSUHFBandTXPower,                                  PeakSatParameters::commsUHFBandTXPower},
+                  {PeakSatParameters::COMMSSBandTXPower,                                    PeakSatParameters::commsSBandTXPower},
+                  {PeakSatParameters::COMMSChannelNumberUHFBand,                            PeakSatParameters::commsChannelNumberUHFBand},
+                  {PeakSatParameters::COMMSChannelNumberSBand,                              PeakSatParameters::commsChannelNumberSBand},
+                  {PeakSatParameters::COMMSLNAGain,                                         PeakSatParameters::commsLNAGain},
+                  {PeakSatParameters::COMMSPAGainUHFBand,                                   PeakSatParameters::commsPAGainUHFBand},
+                  {PeakSatParameters::COMMSPAGainSBand,                                     PeakSatParameters::commsPAGainSBand},
+                  {PeakSatParameters::COMMSVGAGain,                                         PeakSatParameters::commsVGAGain},
+                  {PeakSatParameters::COMMSRSSI,                                            PeakSatParameters::commsRSSI},
+                  {PeakSatParameters::COMMSUHFBandTXOnOff,                                  PeakSatParameters::commsUHFBandTXOnOff},
+                  {PeakSatParameters::COMMSUHFBandRXOnOff,                                  PeakSatParameters::commsUHFBandRXOnOff},
+                  {PeakSatParameters::COMMSSBandTXOnOff,                                    PeakSatParameters::commsSBandTXOnOff},
+                  {PeakSatParameters::COMMSPacketsRejectedCOMMS,                            PeakSatParameters::commsPacketsRejectedCOMMS},
+                  {PeakSatParameters::COMMSInvalidHMAC,                                     PeakSatParameters::commsInvalidHMAC},
+                  {PeakSatParameters::COMMSInvalidPacketStructure,                          PeakSatParameters::commsInvalidPacketStructure},
+                  {PeakSatParameters::COMMSInvalidSpacecraftID,                             PeakSatParameters::commsInvalidSpacecraftID},
+                  {PeakSatParameters::COMMSFrameSequenceCounter,                            PeakSatParameters::commsFrameSequenceCounter},
+                  {PeakSatParameters::COMMSPCBTemperature1,                                 PeakSatParameters::commsPCBTemperature1},
+                  {PeakSatParameters::COMMSPCBTemperature2,                                 PeakSatParameters::commsPCBTemperature2},
+                  {PeakSatParameters::COMMSMCUTemperature,                                  PeakSatParameters::commsMCUTemperature},
+                  {PeakSatParameters::COMMSMCUInputVoltage,                                 PeakSatParameters::commsMCUInputVoltage},
+                  {PeakSatParameters::COMMSMCUBootCounter,                                  PeakSatParameters::commsMCUBootCounter},
+                  {PeakSatParameters::COMMSOnBoardTime,                                     PeakSatParameters::commsOnBoardTime},
+                  {PeakSatParameters::COMMSNANDCurrentlyUsedMemoryPartition,                PeakSatParameters::commsNANDCurrentlyUsedMemoryPartition},
+                  {PeakSatParameters::COMMSLastFailedEvent,                                 PeakSatParameters::commsLastFailedEvent},
+                  {PeakSatParameters::COMMSMCUSystick,                                      PeakSatParameters::commsMCUSystick},
+                  {PeakSatParameters::COMMSFlashInt,                                        PeakSatParameters::commsFlashInt},
+                  {PeakSatParameters::COMMSSRAMInt,                                         PeakSatParameters::commsSRAMInt}};
 }
 
 #endif
